+++ crates/mews_types/src/lib.rs
@@ -48,7 +48,8 @@ pub struct FeedMew {
     pub licks: Vec<AgentPubKey>,
     pub mewmews: Vec<ActionHash>,
     pub deleted_timestamp: Option<Timestamp>,
-    pub author_profile: Profile,
+    pub author_profile: Option<Profile>,
+    pub original_mew: Option<Mew>,
     pub original_mew_author: Option<AgentPubKey>,
     pub original_mew_author_profile: Option<Profile>,
     pub original_mew_deleted_timestamp: Option<Timestamp>,
