+++ dnas/mewsfeed/zomes/coordinator/mews/src/agent_mews.rs
@@ -1,4 +1,4 @@
-use crate::mew::get_mew_with_context;
+use crate::mew_with_context::get_batch_mews_with_context;
 use hdk::prelude::*;
 use mews_integrity::*;
 
@@ -18,12 +18,7 @@ pub fn get_agent_mews(author: AgentPubKey) -> ExternResult<Vec<Record>> {
 pub fn get_agent_mews_with_context(author: AgentPubKey) -> ExternResult<Vec<FeedMew>> {
     let hashes = get_agent_mew_hashes(author)?;
 
-    let feedmews: Vec<FeedMew> = hashes
-        .into_iter()
-        .filter_map(|hash| get_mew_with_context(hash).ok())
-        .collect();
-
-    Ok(feedmews)
+    get_batch_mews_with_context(hashes)
 }
 
 fn get_agent_mew_hashes(author: AgentPubKey) -> ExternResult<Vec<ActionHash>> {
