+++ dnas/mewsfeed/zomes/coordinator/mews/src/hashtag_to_mews.rs
@@ -1,4 +1,4 @@
-use crate::mew::get_mew_with_context;
+use crate::mew_with_context::get_mew_with_context;
 use hdk::prelude::*;
 use mews_integrity::*;
