+++ dnas/mewsfeed/zomes/coordinator/mews/src/lib.rs
@@ -7,6 +7,7 @@ pub mod hashtag_to_mews;
 pub mod licker_to_mews;
 pub mod mention_to_mews;
 pub mod mew;
+pub mod mew_with_context;
 pub mod mew_to_responses;
 pub mod search_tags;
 use mews_integrity::LinkTypes;
