+++ dnas/mewsfeed/zomes/coordinator/mews/src/mention_to_mews.rs
@@ -1,4 +1,4 @@
-use crate::mew::get_mew_with_context;
+use crate::mew_with_context::get_batch_mews_with_context;
 use hdk::prelude::*;
 use mews_integrity::*;
 
@@ -41,12 +41,7 @@ pub fn get_mews_for_mention(mention: AgentPubKey) -> ExternResult<Vec<Record>> {
 pub fn get_mews_for_mention_with_context(mention: AgentPubKey) -> ExternResult<Vec<FeedMew>> {
     let hashes = get_mew_hashes_for_mention(mention)?;
 
-    let feedmews: Vec<FeedMew> = hashes
-        .into_iter()
-        .filter_map(|hash| get_mew_with_context(hash).ok())
-        .collect();
-
-    Ok(feedmews)
+    get_batch_mews_with_context(hashes)
 }
 
 fn get_mew_hashes_for_mention(mention: AgentPubKey) -> ExternResult<Vec<ActionHash>> {
